requirements.txt
google-api-python-client==1.8.3
PyYAML==5.3.1
pandas==0.24.2
unicodecsv==0.14.1
icalendar==4.0.6
#gdata==2.0.18
gdata-python3==3.0.1
oauth2client==4.1.3
#apiclient==1.0.4
paramiko==2.7.1
#bs4==0.0.1
lxml==4.5.2
beautifulsoup4
uuid==1.30

